♻️ Refactor React diagnostics
React Doctorの警告を解消し、不要な再描画と逐次DB読取を削減する。
