+++ src/features/certifications/components/CertificationDrawer.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import { Divider, Drawer, Group, Skeleton, Stack, Switch, Text } from '@mantine/core';
 import { notifications } from '@mantine/notifications';
@@ -33,11 +33,9 @@ type Props = {
  */
 export function CertificationDrawer({ state, onClose }: Props) {
   // 閉じるアニメーション中に表示内容が消えないよう、直近の非 null な state を保持する
-  const [lastState, setLastState] = useState<CertificationDrawerState | null>(null);
+  const [lastState, setLastState] = useState<CertificationDrawerState | null>(state);
 
-  useEffect(() => {
-    if (state) setLastState(state);
-  }, [state]);
+  if (state !== null && state !== lastState) setLastState(state);
 
   const effectiveState = state ?? lastState;
   const isEdit = effectiveState?.mode === 'edit';
