+++ src/features/shift-types/index.ts
@@ -1,4 +1,4 @@
-export { ShiftTypeDrawerContent, type ShiftTypeDrawerState } from './components/ShiftTypeDrawer';
+export type { ShiftTypeDrawerState } from './components/ShiftTypeDrawer';
 export { ShiftTypeList } from './components/ShiftTypeList';
 export { ShiftTypeSettings } from './components/ShiftTypeSettings';
 export type { ShiftTypeListItem } from './schema';
