+++ src/features/users/components/UserDrawer.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import {
   Avatar,
@@ -48,11 +48,9 @@ type Props = {
  */
 export function UserDrawer({ state, onClose }: Props) {
   // 閉じるアニメーション中に表示内容が消えないよう、直近の非 null な state を保持する
-  const [lastState, setLastState] = useState<UserDrawerState | null>(null);
+  const [lastState, setLastState] = useState<UserDrawerState | null>(state);
 
-  useEffect(() => {
-    if (state) setLastState(state);
-  }, [state]);
+  if (state !== null && state !== lastState) setLastState(state);
 
   const effectiveState = state ?? lastState;
